NoitemsView

+++ SwiftUI-Todo-ListApp/Views/ListView.swift
@@ -12,19 +12,27 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-            ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
+        ZStack{
+            if listViewModel.items.isEmpty {
+                Text("No items")
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+            else {
+                List {
+                           ForEach(listViewModel.items) { item in
+                           ListRowView(item: item)
+                                   .onTapGesture {
+                                       withAnimation(.linear) {
+                                           listViewModel.updateItem(item: item)
+                                       }
+                                   }
+                           }
+                           .onDelete(perform: listViewModel.deleteItem)
+                           .onMove(perform: listViewModel.moveItem)
+                       }
+                       .listStyle(PlainListStyle())
+            }
         }
-        .listStyle(PlainListStyle())
+       
         .navigationTitle("Todo List 📝")
         .navigationBarItems(leading: EditButton(), trailing:
         NavigationLink("Add", destination: AddView())
